+++ apiosintDS/modules/dosearch.py
@@ -18,9 +18,10 @@ class dosearch():
     def __init__(self, lists, verbose):
         self.lists = lists
         self.verbose = verbose
+        self.hashlookup = False
 
         ### !!! ### o_O
-        self.entitype = ['url', 'ip', 'domain'] #Supported Entities type
+        self.entitype = ['url', 'ip', 'domain', 'hash'] #Supported Entities type
         self.output = {
                         "url": {
                                     "items": [],
@@ -58,20 +59,38 @@ class dosearch():
                                                 "url": "undefined"
                                     }
                                },
+                        "hash": {
+                                    "items": [],
+                                    "statistics": {
+                                                "itemFound": 0,
+                                                "itemSubmitted": 0
+                                    },
+                                    "list": {
+                                                "file": "undefined",
+                                                "date": "undefined",
+                                                "url": "undefined"
+                                    }
+                               }, 
                       }
         self.getResults = self.ExecSearch()
 
-    def goodResult(self, item, entype):
+    def goodResult(self, item, entype, relatedByHash=False, hashes=False):
         ret = {}
         ret["item"] = item
         ret["response"] = True
         ret["response_text"] = "Item found in "+self.lists["lookup"][entype]["file"]+" list"
-        if entype != 'url':
+        if entype in ['ip', 'domain']:
             ret["related_urls"] = self.parseRelatedUrl(item)
-        else:
+        elif entype == 'url':
+            searchUrlHash = self.searchUrlHash(item)
+            ret["hashes"] = searchUrlHash
             getItem = urlparse(item)
-            ret["related_urls"] = self.parseRelatedUrl(getItem.hostname)
-            ret["related_urls"].remove(item)
+            ret["related_urls"] = self.parseRelatedUrl(getItem.hostname, item)
+        elif entype == 'hash':
+            ret["hashes"] = hashes
+            ret["related_urls"] = relatedByHash
+        else:
+            pass
         return ret
 
     def badResult(self, item, entype):
@@ -79,31 +98,58 @@ class dosearch():
         ret["item"] = item
         ret["response"] = False
         ret["response_text"] = "Item not found"
-        ret["related_urls"] = []
+        ret["related_urls"] = ""
         return ret
 
-    def parseRelatedUrl(self, item):
+    def parseRelatedUrl(self, item, ItemURL=False):
         relatedURLs = []
         resUrls = self.lists["lookup"]["url"]["items"]
         for lineurl in resUrls:
             ParseURL = urlparse(lineurl)
-            if ParseURL.hostname == item:
-                relatedURLs.append(lineurl)
+            if (ParseURL.hostname == item) and (lineurl != ItemURL):
+            
+                relhashes = self.searchUrlHash(lineurl)
+                
+                reldict = {'url': lineurl, 'hashes': relhashes}
+                relatedURLs.append(reldict)
         return relatedURLs
 
+    def searchUrlHash(self, item):
+        ret = False
+        for obj in self.lists["lookup"]["hash"]["items"]:
+            if item in self.lists["lookup"]["hash"]["items"][obj]["url"]:
+                return {"md5": self.lists["lookup"]["hash"]["items"][obj]["md5"], "sha1": self.lists["lookup"]["hash"]["items"][obj]["sha1"], "sha256": self.lists["lookup"]["hash"]["items"][obj]["sha256"]}
+         
+        return ret  
+            
+    def searchHash(self, item):
+        hashtype_dict = {32: "md5", 40: "sha1", 64: "sha256"}
+        hashlen = len(item)
+        hashtype = hashtype_dict[hashlen]
+        for obj in self.lists["lookup"]["hash"]["items"]:
+            if item in self.lists["lookup"]["hash"]["items"][obj][hashtype]:
+                hashes = {"md5": self.lists["lookup"]["hash"]["items"][obj]["md5"], "sha1": self.lists["lookup"]["hash"]["items"][obj]["sha1"], "sha256": self.lists["lookup"]["hash"]["items"][obj]["sha256"]}
+                self.output["hash"]["items"].append(self.goodResult(item, "hash", self.lists["lookup"]["hash"]["items"][obj]["url"], hashes))
+                self.output["hash"]["statistics"]["itemFound"] +=1
+                break
+                
     def ExecSearch(self):
         for entype in self.entitype:
             if len(self.lists["input"]["entities"][entype]) > 0:
                 self.output[entype]["list"]["date"] = self.lists["lookup"][entype]["date"]
                 self.output[entype]["list"]["url"] = self.lists["lookup"][entype]["url"]
                 self.output[entype]["list"]["file"] = self.lists["lookup"][entype]["file"]
+                
                 for item in self.lists["input"]["entities"][entype]:
-                    if item in self.lists["lookup"][entype]["items"]:
-                        self.output[entype]["items"].append(self.goodResult(item, entype))
-                        self.output[entype]["statistics"]["itemFound"] +=1
+                    if entype == "hash":
+                        self.searchHash(item)
                     else:
-                        if self.verbose:
-                            self.output[entype]["items"].append(self.badResult(item, entype))
+                        if item in self.lists["lookup"][entype]["items"]:
+                            self.output[entype]["items"].append(self.goodResult(item, entype))
+                            self.output[entype]["statistics"]["itemFound"] +=1
+                        else:
+                            if self.verbose:
+                                self.output[entype]["items"].append(self.badResult(item, entype))
                 self.output[entype]["statistics"]["itemSubmitted"] = len(self.lists["input"]["entities"][entype])
             else:
                 del self.output[entype]