+++ apiosintDS/modules/listutils.py
@@ -6,6 +6,7 @@ import validators
 import logging
 import json
 from datetime import datetime
+import re
 import pytz
 
 italyTZ = pytz.timezone("Europe/Rome")
@@ -25,7 +26,7 @@ class listutils():
         self.clearcache = clearcache
 
         ### !!! ### o_O
-        self.template = {"url": [], "ip": [], "domain": []}
+        self.template = {"url": [], "ip": [], "domain": [], "hash": []}
         self.items = self.getItems()
         self.entities = dict(self.getEntities())
         self.checkdate = italyTZ.localize(datetime.strptime(datetime.today().strftime('%Y-%m-%d %H:%M:%S'), '%Y-%m-%d %H:%M:%S'))
@@ -39,6 +40,18 @@ class listutils():
             self.items = self.listfile
         return self.items
 
+    def validatehash(self, hstring):
+        ret = False
+        digits = len(hstring)
+        if digits in [32, 40, 64]:
+            try:
+                ret = re.match(r'^[0-9a-f]{'+str(digits)+'}$', hstring).group(0)
+            except:
+                logging.warning("Invalid hash detected: failed validation for ["+hstring+"]")
+        else:
+            logging.warning("Invalid hash detected: failed lenght check for ["+hstring+" len("+str(digits)+")]")
+        return ret
+
     def getEntities(self):
         counter = 0
         for line in self.items:
@@ -51,25 +64,39 @@ class listutils():
             elif validators.domain(line):
                 counter +=1
                 self.template["domain"].append(line)
+            elif self.validatehash(line):
+                counter +=1
+                self.template["hash"].append(line)
             else:
-                logging.warning("{} is not a valid IPv4/domain/url. REMOVED!.".format(line))
+                logging.warning("{} is not a valid IPv4/domain/url/hash. REMOVED!.".format(line))
         if counter == 0:
-            logging.error("No valid elements detected, sorry! Supported entities are IPv4/Domain/URL.")
+            logging.error("No valid elements detected, sorry! Supported entities are IPv4/Domain/URL/Hash ['MD5', 'SHA1', 'SHA256'].")
             exit(1)
         ret = {}
         ret["entities"] = self.template
-        ret["centities"] = {"url": len(self.template["url"]), "ip": len(self.template["ip"]), "domain": len(self.template["domain"])}
+        ret["centities"] = {"url": len(self.template["url"]), "ip": len(self.template["ip"]), "domain": len(self.template["domain"]), "hash": len(self.template["hash"])}
         return ret
 
-    def getListDate(self, context):
-        return italyTZ.localize(datetime.strptime(context[9][16:35], '%Y-%m-%d %H:%M:%S'))
+    def getListDate(self, context, hashes=False):
+        if hashes:
+            context = json.loads(context)
+            listdate = context["generatedAt"]
+        else:
+            listdate = context[9][16:35]
+        return italyTZ.localize(datetime.strptime(listdate, '%Y-%m-%d %H:%M:%S'))
 
-    def getListItems(self, context):
-        return context[11:]
+    def getListItems(self, context, hashes=False):
+        if hashes:
+            context = json.loads(context)
+            listitems = context["lookup"]
+        else:
+            listitems = context[11:]
+        return listitems
 
     def getCache(self): #Mmhhh, not a good method at all, but works. Added to TODO list for early refactor. 2019-10-12
         cached = {}
         for entity in self.entities["centities"]:
+            getHash = (True if entity == "hash" else False)
             cached[entity] = {}
             cached[entity]["file"] = "latest"+entity+"s.txt"
             if (self.entities["centities"][entity] == 0) and (entity in ["ip", "domain"]):
@@ -79,33 +106,39 @@ class listutils():
                     cachedfile = self.cachedir+cached[entity]["file"]
                     if os.path.exists(cachedfile):
                         if self.clearcache:
-                            dwlist = self.downloadLists(entity)
+                            dwlist = self.downloadLists(entity, getHash)
                             cached[entity]["date"] = self.getListDate(dwlist["text"])
                             cached[entity]["items"] = self.getListItems(dwlist["text"])
                             cached[entity]["url"] = dwlist["url"]
                         else:
-                            dwlist = [line.rstrip('\n') for line in open(cachedfile)]
-                            listdate = self.getListDate(dwlist)
-                            listitems = self.getListItems(dwlist)
+                            if entity == "hash":
+                                dwlist = json.loads(cacheddownloadListsfile)
+                            else:
+                                dwlist = [line.rstrip('\n') for line in open(cachedfile)]
+                            
+                            listdate = self.getListDate(dwlist, getHash)
+                            listitems = self.getListItems(dwlist, getHash)
+                                
                             diffdate = ((self.checkdate-listdate).total_seconds())/3600
+                            
                             if ((diffdate) < 4) and (len(listitems) > 0):
                                 cached[entity]["date"] = listdate
                                 cached[entity]["items"] = listitems
                                 cached[entity]["url"] = "Loaded from cache '"+cachedfile+"'"
                             else:
-                                dwlist = self.downloadLists(entity)
-                                cached[entity]["date"] = self.getListDate(dwlist["text"])
-                                cached[entity]["items"] = self.getListItems(dwlist["text"])
+                                dwlist = self.downloadLists(entity, getHash)
+                                cached[entity]["date"] = self.getListDate(dwlist["text"], getHash)
+                                cached[entity]["items"] = self.getListItems(dwlist["text"], getHash)
                                 cached[entity]["url"] = dwlist["url"]
                     else:
-                        dwlist = self.downloadLists(entity)
-                        cached[entity]["date"] = self.getListDate(dwlist["text"])
-                        cached[entity]["items"] = self.getListItems(dwlist["text"])
+                        dwlist = self.downloadLists(entity, getHash)
+                        cached[entity]["date"] = self.getListDate(dwlist["text"], getHash)
+                        cached[entity]["items"] = self.getListItems(dwlist["text"], getHash)
                         cached[entity]["url"] = dwlist["url"]
                 else:
-                    dwlist = self.downloadLists(entity)
-                    cached[entity]["date"] = self.getListDate(dwlist["text"])
-                    cached[entity]["items"] = self.getListItems(dwlist["text"])
+                    dwlist = self.downloadLists(entity, getHash)
+                    cached[entity]["date"] = self.getListDate(dwlist["text"], getHash)
+                    cached[entity]["items"] = self.getListItems(dwlist["text"], getHash)
                     cached[entity]["url"] = dwlist["url"]
                 cached[entity]["date"] = str(cached[entity]["date"])
         return cached
@@ -121,7 +154,9 @@ class listutils():
             logging.error("Retry without -c, --cache option.")
             exit(1)
 
-    def downloadLists(self, entity):
+    def downloadLists(self, entity, hashes=False):
+        # micro patch for hashes vs hash
+        entity = entity+"e" if entity == "hash" else entity
         ret = {}
         ret["url"] = self.urls['master_url']+entity+'s.txt'
         r = requests.get(ret["url"])
@@ -142,7 +177,10 @@ class listutils():
             logging.error("The downloaded list seems to be empty!\n")
             logging.error(self.status_error(entity))
             exit(1)
-        return_text = text.split('\n')
+        if hashes:
+            return_text = text
+        else:
+            return_text = text.split('\n')
         if len(return_text[11:]) == 0: #just because I'm a very paranoid man, but same time a lazy one ^_^'''
             logging.error("The downloaded list seems to be empty!\n")
             logging.error(self.status_error(entity))